Unreviewed debug build fix after r103115.
* dom/Document.cpp:
(WebCore::Document::cachedCollection):


Canonical link: https://commits.webkit.org/91486@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@103122 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebCore/dom/Document.cpp
@@ -4192,7 +4192,7 @@ bool Document::hasSVGRootNode() const
 
 const RefPtr<HTMLCollection>& Document::cachedCollection(CollectionType type)
 {
-    ASSERT(type < NumUnnamedDocumentCachedTypes);
+    ASSERT(static_cast<unsigned>(type) < NumUnnamedDocumentCachedTypes);
     if (!m_collections[type])
         m_collections[type] = HTMLCollection::createForCachingOnDocument(this, type);
     return m_collections[type];
